AUTO WIFI SERVICE
this project is all about making raspberry pi to scan and connect to all available hotspot available. And it sent all the traffic to router that is connected to the lan port . Thereby acting like a repeater we also have installed pihole for ads blocking and also added VPN for private network connection.